// src/app/generated/models/supervision-visit.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface SupervisionVisit {
  createdDateTime?: string;
  dateOfVisit?: string;
  hoursCount?: number;
  id?: number;
  lastModifiedDateTime?: string;
  payable?: boolean;
  supervisionId?: number;
}